+++ app/Filament/Resources/RumbleCampaignDataResource/Pages/CreateRumbleCampaignData.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\RumbleCampaignDataResource\Pages;
+
+use App\Filament\Resources\RumbleCampaignDataResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateRumbleCampaignData extends CreateRecord
+{
+    protected static string $resource = RumbleCampaignDataResource::class;
+}