+++ app/Filament/Resources/RumbleCampaignDataResource/Pages/EditRumbleCampaignData.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\RumbleCampaignDataResource\Pages;
+
+use App\Filament\Resources\RumbleCampaignDataResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditRumbleCampaignData extends EditRecord
+{
+    protected static string $resource = RumbleCampaignDataResource::class;
+}