+++ app/Filament/Resources/RumbleCampaignDataResource/Pages/ListRumbleCampaignData.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\RumbleCampaignDataResource\Pages;
+
+use App\Filament\Resources\RumbleCampaignDataResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListRumbleCampaignData extends ListRecords
+{
+    protected static string $resource = RumbleCampaignDataResource::class;
+}