+++ app/Filament/Resources/RumbleCampaignDataResource/Pages/ViewRumbleCampaignData.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\RumbleCampaignDataResource\Pages;
+
+use App\Filament\Resources\RumbleCampaignDataResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewRumbleCampaignData extends ViewRecord
+{
+    protected static string $resource = RumbleCampaignDataResource::class;
+}